+++ gamification-specs/src/test/java/ch/heigvd/gamification/api/spec/steps/ApplicationSteps.java
@@ -1,25 +1,18 @@
 package ch.heigvd.gamification.api.spec.steps;
 
 import ch.heigvd.gamification.ApiException;
-import ch.heigvd.gamification.ApiResponse;
 import ch.heigvd.gamification.api.DefaultApi;
 import ch.heigvd.gamification.api.dto.ApiKey;
 import ch.heigvd.gamification.api.dto.Application;
-import ch.heigvd.gamification.api.dto.Badge;
-import ch.heigvd.gamification.api.dto.Event;
-import ch.heigvd.gamification.api.dto.EventEventparams;
-import ch.heigvd.gamification.api.dto.User;
 import ch.heigvd.gamification.api.spec.helpers.Environment;
-import io.cucumber.java.en.And;
+
 import io.cucumber.java.en.Given;
-import io.cucumber.java.en.Then;
+
 import io.cucumber.java.en.When;
 
-import java.util.List;
+
 import java.util.UUID;
 
-import static org.junit.Assert.assertEquals;
-import static org.junit.Assert.assertNotNull;
 
 public class ApplicationSteps {
 
@@ -27,14 +20,18 @@ public class ApplicationSteps {
     private DefaultApi api;
     private BasicSteps basicSteps;
 
-    Application application;
-    UUID apiKey;
+    private Application application;
+    private UUID apiKey;
 
 
     public ApplicationSteps(Environment environment, BasicSteps basicSteps) {
         this.environment = environment;
         this.api = environment.getApi();
         this.basicSteps=basicSteps;
+    }
+
+    public UUID getApiKey() {
+        return apiKey;
     }
 
     @Given("I have an application payload")
